Moved already present EXISTS tests
