+++ binding/rdf4j/src/test/java/it/unibz/inf/ontop/rdf4j/repository/ExistsTest.java
@@ -1,11 +1,14 @@
 package it.unibz.inf.ontop.rdf4j.repository;
 
 import com.google.common.collect.ImmutableList;
+import com.google.common.collect.ImmutableSet;
 import org.eclipse.rdf4j.query.QueryEvaluationException;
 import org.junit.AfterClass;
 import org.junit.Assert;
 import org.junit.BeforeClass;
 import org.junit.Test;
+
+import static org.junit.Assert.assertEquals;
 
 public class ExistsTest extends AbstractRDF4JTest {
     private static final String OBDA_FILE = "/person/person_exists.obda";
@@ -29,20 +32,23 @@ public class ExistsTest extends AbstractRDF4JTest {
                 "   FILTER EXISTS { ?v :firstName ?fname }\n" +
                 "} ORDER BY ?v \n";
 
-        runQueryAndCompare(sparql, ImmutableList.of("http://person.example.org/person/1", "http://person.example.org/person/2", "http://person.example.org/person/5"));
+        runQueryAndCompare(sparql, ImmutableList.of("http://person.example.org/person/1",
+                "http://person.example.org/person/2", "http://person.example.org/person/5"));
     }
 
     @Test
-    public void testFilterConjunction() {
+    public void testFilterDisjunction() {
         String sparql = "PREFIX  : <http://person.example.org/> \n" +
                 "SELECT ?v WHERE {\n" +
                 "   ?v a :Person .\n" +
                 "   FILTER (?lname = \"Smith\" || EXISTS { ?v :firstName ?fname })\n" +
                 "} ORDER BY ?v \n";
 
-        runQueryAndCompare(sparql, ImmutableList.of("http://person.example.org/person/1", "http://person.example.org/person/2", "http://person.example.org/person/5"));
+        runQueryAndCompare(sparql, ImmutableList.of("http://person.example.org/person/1",
+                "http://person.example.org/person/2", "http://person.example.org/person/5"));
     }
 
+    // Non overlapping variables are not supported
     @Test(expected = QueryEvaluationException.class)
     public void testFilterNoCommonVariables() {
         String sparql = "PREFIX  : <http://person.example.org/> \n" +
@@ -95,9 +101,11 @@ public class ExistsTest extends AbstractRDF4JTest {
                 "       FILTER EXISTS {?v :firstName ?fname }\n" +
                 "   }\n" +
                 "} ORDER BY ?v";
-        runQueryAndCompare(sparql, ImmutableList.of("http://person.example.org/person/1", "http://person.example.org/person/2", "http://person.example.org/person/5"));
+        runQueryAndCompare(sparql, ImmutableList.of("http://person.example.org/person/1",
+                "http://person.example.org/person/2", "http://person.example.org/person/5"));
     }
 
+    // Not supported (no variables in the exists graph pattern)
     @Test(expected = QueryEvaluationException.class)
     public void testOnlyConstantsInFilter() {
         String sparql = "PREFIX : <http://person.example.org/>\n" +
@@ -108,6 +116,7 @@ public class ExistsTest extends AbstractRDF4JTest {
         Assert.assertEquals(5, runQueryAndCount(sparql));
     }
 
+    // Not supported (no variables in the exists graph pattern)
     @Test(expected = QueryEvaluationException.class)
     public void testOnlyConstantsInFilter1() {
         String sparql = "PREFIX : <http://person.example.org/>\n" +
@@ -139,6 +148,112 @@ public class ExistsTest extends AbstractRDF4JTest {
                 "} ORDER BY ?v \n";
 
         runQueryAndCompare(sparql, ImmutableList.of("http://person.example.org/person/1"));
+    }
+
+    @Test
+    public void testNotExists() {
+        String sparql = "PREFIX  : <http://person.example.org/> \n" +
+                "SELECT ?v WHERE {\n" +
+                "    ?v a :Person .\n" +
+                "    FILTER NOT EXISTS { ?v :firstName ?fname }\n" +
+                "}\n";
+
+        runQueryAndCompare(sparql, ImmutableSet.of("http://person.example.org/person/3",
+                "http://person.example.org/person/4", "http://person.example.org/person/6"));
+    }
+
+    // Non overlapping variables are not supported
+    @Test(expected = QueryEvaluationException.class)
+    public void testNonOverlappingVariables() {
+        String sparql = "SELECT * \n" +
+                "WHERE { ?s ?p ?o \n" +
+                "   FILTER NOT EXISTS { ?x ?y ?z } \n" +
+                "}\n";
+        int countResults = runQueryAndCount(sparql);
+        assertEquals(0, countResults);
+    }
+
+    @Test
+    public void testFilterNotExistsConstants1() {
+        String sparql = "PREFIX : <http://person.example.org/>\n" +
+                "SELECT ?v WHERE { " +
+                "   ?v :firstName ?fname ; \n" +
+                "   FILTER NOT EXISTS { <http://person.example.org/person/1> :firstName ?fname } \n" +
+                "}\n";
+        runQueryAndCompare(sparql, ImmutableSet.of("http://person.example.org/person/2"));
+    }
+
+    // Not supported (no variables present in the not exists graph pattern)
+    @Test(expected = QueryEvaluationException.class)
+    public void testFilterNotExistsAllConstants() {
+        String sparql = "PREFIX : <http://person.example.org/>\n" +
+                "SELECT ?v WHERE { " +
+                "       ?v :firstName ?fname ; \n" +
+                "       FILTER NOT EXISTS { <http://person.example.org/person/1> :firstName \"Roger\" } \n" +
+                "}\n";
+        int countResults = runQueryAndCount(sparql);
+        assertEquals(0, countResults);
+    }
+
+    // The inner filter variables not bound in the not exists graph pattern is not supported
+    @Test(expected = QueryEvaluationException.class)
+    public void testFilterOnUnboundVariable() {
+        String sparql = "PREFIX : <http://person.example.org/>\n" +
+                "SELECT ?v WHERE {\n" +
+                "        ?v :firstName ?fname .\n" +
+                "        FILTER NOT EXISTS {\n" +
+                "                ?v :nickname ?nick .\n" +
+                "                FILTER(?fname = ?nick)\n" +
+                "        }\n" +
+                "}";
+        runQueryAndCompare(sparql, ImmutableSet.of("http://person.example.org/person/1"));
+    }
+
+    @Test
+    public void testFilterSharedConstants() {
+        String sparql = "PREFIX : <http://person.example.org/>\n" +
+                "SELECT ?v WHERE {\n" +
+                "    ?v a :Person ;\n" +
+                "        :firstName \"Roger\" .\n" +
+                "   FILTER NOT EXISTS {\n" +
+                "       ?v :firstName ?fname .\n" +
+                "       FILTER (?fname NOT IN (\"Roger\", \"Paul\"))\n" +
+                "    }\n" +
+                "}\n";
+        runQueryAndCompare(sparql, ImmutableSet.of("http://person.example.org/person/1", "http://person.example.org/person/5"));
+    }
+
+    @Test
+    public void testGraphPatternFilterOrder() {
+        String sparql = "PREFIX : <http://person.example.org/>\n" +
+                "SELECT ?v ?lname WHERE {\n" +
+                "    ?v a :Person .\n" +
+                "    FILTER NOT EXISTS { ?v :firstName ?fname }\n" +
+                "    ?v :lastName ?lname\n" +
+                "}\n";
+        runQueryAndCompare(sparql, ImmutableSet.of("http://person.example.org/person/3"));
+    }
+
+    @Test
+    public void testNotExistsSharedVariables() {
+        String sparql = "PREFIX : <http://person.example.org/>\n" +
+                "SELECT ?v WHERE {\n" +
+                "    ?v a :Person .\n" +
+                "    FILTER NOT EXISTS { ?person :nickname ?sharedName }\n" +
+                "    ?v :firstName ?sharedName\n" +
+                "} ORDER BY ?v\n";
+        runQueryAndCompare(sparql, ImmutableList.of("http://person.example.org/person/1", "http://person.example.org/person/5"));
+    }
+
+    @Test
+    public void testExistsSharedVariables() {
+        String sparql = "PREFIX : <http://person.example.org/>\n" +
+                "SELECT ?v WHERE {\n" +
+                "    ?v a :Person .\n" +
+                "    FILTER EXISTS { ?person :nickname ?sharedName }\n" +
+                "    ?v :firstName ?sharedName\n" +
+                "} ORDER BY ?v\n";
+        runQueryAndCompare(sparql, ImmutableList.of("http://person.example.org/person/2"));
     }
 
     @Test
@@ -187,6 +302,19 @@ public class ExistsTest extends AbstractRDF4JTest {
         runQueryAndCompare(sparql, ImmutableList.of("http://person.example.org/person/4", "http://person.example.org/person/6"));
     }
 
+    @Test
+    public void testExistsWithCoalesce() {
+        String sparql = "PREFIX  : <http://person.example.org/> \n" +
+                "SELECT ?v WHERE {\n" +
+                "   ?v a :Person .\n" +
+                "   FILTER(COALESCE(EXISTS { ?v :firstName ?fname }, false))\n" +
+                "} ORDER BY ?v \n";
+
+        runQueryAndCompare(sparql, ImmutableList.of("http://person.example.org/person/1", "http://person.example.org/person/2",
+                "http://person.example.org/person/5"));
+    }
+
+    // exists inside aggregation not supported
     @Test(expected = QueryEvaluationException.class)
     public void testExistsWithAggregation() {
         String sparql = "PREFIX  : <http://person.example.org/> \n" +
@@ -199,6 +327,7 @@ public class ExistsTest extends AbstractRDF4JTest {
         runQueryAndCompare(sparql, ImmutableList.of("1", "1", "0", "0", "1", "0"));
     }
 
+    // exists inside aggregation not yet supported
     @Test(expected = QueryEvaluationException.class)
     public void testExistsWithAggregationNoSharedVars() {
         String sparql = "PREFIX  : <http://person.example.org/> \n" +
@@ -211,6 +340,7 @@ public class ExistsTest extends AbstractRDF4JTest {
         runQueryAndCompare(sparql, ImmutableList.of("1", "1", "1", "1", "1", "1"));
     }
 
+    // exists inside order by not yet supported
     @Test(expected = QueryEvaluationException.class)
     public void testExistsWithOrderBy() {
         String sparql = "PREFIX  : <http://person.example.org/> \n" +
@@ -222,6 +352,7 @@ public class ExistsTest extends AbstractRDF4JTest {
                 "http://person.example.org/person/3", "http://person.example.org/person/4", "http://person.example.org/person/6"));
     }
 
+    // exists inside order by not yet supported
     @Test(expected = QueryEvaluationException.class)
     public void testExistsWithOrderByNoSharedVars() {
         String sparql = "PREFIX  : <http://person.example.org/> \n" +
@@ -231,5 +362,17 @@ public class ExistsTest extends AbstractRDF4JTest {
 
         runQueryAndCompare(sparql, ImmutableList.of("http://person.example.org/person/1", "http://person.example.org/person/2", "http://person.example.org/person/3",
                 "http://person.example.org/person/4", "http://person.example.org/person/5", "http://person.example.org/person/6"));
+    }
+
+    // Not yet supported
+    @Test(expected = QueryEvaluationException.class)
+    public void testNullableOrTheRight() {
+        String sparql = "PREFIX : <http://person.example.org/>\n" +
+                "SELECT ?v WHERE {\n" +
+                "    ?v a :Person .\n" +
+                "    FILTER NOT EXISTS { ?person a :Person . OPTIONAL { ?person :locality ?sharedName } }\n" +
+                "    ?v :nickname ?sharedName\n" +
+                "}\n";
+        runQueryAndCompare(sparql, ImmutableSet.of());
     }
 }
