--- binding/rdf4j/src/test/resources/person/person_not_exists.obda
@@ -1,17 +1,0 @@
-[PrefixDeclaration]
-:		http://person.example.org/
-owl:		http://www.w3.org/2002/07/owl#
-rdf:		http://www.w3.org/1999/02/22-rdf-syntax-ns#
-xml:		http://www.w3.org/XML/1998/namespace
-xsd:		http://www.w3.org/2001/XMLSchema#
-obda:		https://w3id.org/obda/vocabulary#
-rdfs:		http://www.w3.org/2000/01/rdf-schema#
-quest:		http://obda.org/quest#
-
-[MappingDeclaration] @collection [[
-mappingId	MAPID-person
-target		:person/{id} a :Person ; :firstName {firstName} ; :lastName {lastName} ; :nickname {nickname} ; :country {country} ; :locality {locality} .
-source		SELECT "id", "firstName", "lastName", "nickname", "country", "locality" FROM "person";
-
-]]
-
