--- binding/rdf4j/src/test/resources/person/person_not_exists.sql
@@ -1,22 +1,0 @@
-create table "person" (
-                          "id" int primary key,
-                          "firstName" varchar(100),
-                          "lastName" varchar(100),
-                          "nickname" varchar(100),
-                          "status" int NOT NULL,
-                          "country" varchar(100) NOT NULL,
-                          "locality" varchar(100)
-);
-
-insert into "person" ("id", "firstName", "lastName", "nickname", "status", "country", "locality") values (1, 'Roger', 'Smith', 'Roger Smith', 1, 'it', NULL);
-insert into "person" ("id", "firstName", "lastName", "nickname", "status", "country", "locality") values (2, 'John', NULL, 'John',0, 'uk', 'Belfast');
-insert into "person" ("id", "firstName", "lastName", "nickname", "status", "country", "locality") values (3, NULL, 'Doe', 'Jane Doe', 0, 'ie', 'Dublin');
-insert into "person" ("id", "firstName", "lastName", "nickname", "status", "country", "locality") values (4, NULL, NULL, 'Corto Maltese', 0, 'mt', 'Valletta');
-
-create table "statuses" (
-                            "status_id" int primary key,
-                            "description" varchar(100) NOT NULL
-);
-
-insert into "statuses" ("status_id", "description") values (1, 'OK');
-
